+++ App.tsx
@@ -5,7 +5,6 @@
  * @format
  */
 
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import React from 'react';
@@ -13,6 +12,7 @@ import { StatusBar } from 'react-native';
 import { PaperProvider } from 'react-native-paper';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { Provider } from 'react-redux';
+import { navigationRef } from './src/navigation';
 import { HomeTabNav } from './src/navigation/MainMenuTabNav';
 import { RootStackParamList } from './src/navigation/types';
 import { store } from './src/redux/store';
@@ -20,7 +20,6 @@ import { LoginScreen } from './src/screens/Login';
 import { LemmyDarkTheme } from './src/theme';
 import { ScreenNames } from './src/types';
 
-const Tab = createBottomTabNavigator();
 const Stack = createNativeStackNavigator<RootStackParamList>();
 
 function App(): JSX.Element {
@@ -28,12 +27,13 @@ function App(): JSX.Element {
     <Provider store={store}>
       <PaperProvider theme={LemmyDarkTheme}>
         <SafeAreaProvider>
-          <NavigationContainer theme={LemmyDarkTheme}>
+          <NavigationContainer theme={LemmyDarkTheme} ref={navigationRef}>
             <StatusBar
               barStyle="light-content"
               backgroundColor={LemmyDarkTheme.colors.background}
             />
             <Stack.Navigator>
+              {/* Tabs */}
               <Stack.Screen
                 name="Tabs"
                 component={HomeTabNav}
@@ -41,7 +41,14 @@ function App(): JSX.Element {
                   headerShown: false,
                 }}
               />
-              <Stack.Screen name={ScreenNames.Login} component={LoginScreen} />
+              {/* Login */}
+              <Stack.Screen
+                name={ScreenNames.Login}
+                component={LoginScreen}
+                // options={{
+                //   presentation: 'fullScreenModal',
+                // }}
+              />
             </Stack.Navigator>
             {/* <Tab.Navigator> */}
             {/* Feed */}