+++ src/components/ErrorMsg.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { StyleSheet, View } from 'react-native';
+import { StyleProp, StyleSheet, View, ViewStyle } from 'react-native';
 import { Text } from 'react-native-paper';
 import { IError } from '../types';
 
@@ -7,17 +7,20 @@ import { IError } from '../types';
 export interface IErrorMsgProps {
   /** The error object */
   error: IError;
+  containerStyle?: StyleProp<ViewStyle>;
 }
 
 /**
  * A simple component that can display an error message and code
  * @param props - The {@link IError} to display
  */
-export const ErrorMsg = (props: IError): JSX.Element => {
+export const ErrorMsg = (props: IErrorMsgProps): JSX.Element => {
   const styles = createStyleSheet();
   return (
-    <View style={styles.container}>
-      <Text style={styles.message}>{props.message}</Text>
+    <View style={props.containerStyle}>
+      <View style={styles.container}>
+        <Text style={styles.message}>{props.error.message}</Text>
+      </View>
     </View>
   );
 };
@@ -27,6 +30,7 @@ const createStyleSheet = () => {
     container: {
       backgroundColor: '#FF0000',
       padding: 14,
+      borderRadius: 12,
     },
     message: {
       color: '#FFFFFF',