+++ src/components/ListFooterLoading.tsx
@@ -5,6 +5,7 @@ import { getCommunityListItems } from '../features/communities/selectors';
 import { getLemmyAPILoading } from '../features/lemmy/selectors';
 import { useAppSelector } from '../redux/hooks';
 
+/** Loader view shown at the bottom of Lists when they load more data */
 export const ListFooterLoading = (): JSX.Element => {
   const loading = useAppSelector(getLemmyAPILoading);
   const communities = useAppSelector(getCommunityListItems);
@@ -12,7 +13,7 @@ export const ListFooterLoading = (): JSX.Element => {
   return loading && communities.length !== 0 ? (
     <View style={styles.container}>
       <Text>Loading...</Text>
-      <ActivityIndicator />
+      <ActivityIndicator style={styles.loadingIndicator} />
     </View>
   ) : (
     <View />
@@ -25,6 +26,10 @@ const createStyleSheet = () => {
       padding: 12,
       justifyContent: 'center',
       alignItems: 'center',
+    },
+    loadingIndicator: {
+      marginTop: 8,
+      marginBottom: 12,
     },
   });
 };