+++ src/features/lemmy/types.ts
@@ -5,19 +5,31 @@ import { IError } from '../../types';
 export interface ILemmyState {
   /** Error object for any of the API interactions */
   error?: IError;
+  loginError?: IError;
   /** `true` if we're loading/sending data */
   loading: boolean;
   /** The response from a successful login */
   loginResponse?: LoginResponse;
 }
 
-/** The error object returned if the {@link lemmyLogin} action fails */
+export interface ILemmyInstance {
+  /**
+   * The instance name not formatted as a url. ie `'lemmy.ml'`
+   */
+  name: string;
+  /**
+   * The instance formatted as a url. ie `'https://lemmy.ml'`
+   */
+  href: string;
+}
+
+/** The error object returned if the {@link features/lemmy/actions!lemmyLogin} action fails */
 export interface ILemmyLoginError {
   /** The error message */
   error: string;
 }
 
-/** Params for the {@link lemmyLogin} action */
+/** Params for the {@link features/lemmy/actions!lemmyLogin} action */
 export interface ILemmyLoginParams {
   /** The param for the login api call */
   loginForm: Login;
@@ -25,10 +37,22 @@ export interface ILemmyLoginParams {
   instanceUrl: string;
 }
 
+// ---- API Responses
+
+export interface ILemmyLoginErrorResponse {
+  error: string;
+}
+
 /** Error messages related to logging in */
-export enum LoginErrors {
+export enum LemmyLoginErrors {
   /** A possible response if the user/pass is wrong. */
-  UserOrPassInvalid = 'couldnt_find_that_username_or_email',
+  UserOrPassInvalid = 'Error: couldnt_find_that_username_or_email',
+  PasswordIncorrect = 'Error: password_incorrect',
+}
+
+export enum LemmyErrorMsgs {
+  UserOrPassInvalid = 'There was a problem with your username or password. Please verify them and try again.',
+  UnknownError = 'An unknown error occured.',
 }
 
 export enum LemmyAPIPaths {