+++ src/navigation/MainMenuTabNav.tsx
@@ -1,23 +1,24 @@
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import MaterialIcon from 'react-native-vector-icons/MaterialCommunityIcons';
+import { CommunityListHeader } from '../components/communities/CommunityListHeader';
+import { AccountSwitcher } from '../components/profile/AccountSwitcher';
 import { CommunitiesScreen } from '../screens/Communities';
-import { LoginScreen } from '../screens/Login';
+import { MainMenuScreen } from '../screens/MainMenu';
 import { LemmyDarkTheme } from '../theme';
 import { MaterialIconNames, ScreenNames, TAB_ICON_SIZE } from '../types';
-import { CommunityListHeader } from '../components/communities/CommunityListHeader';
 
 const Tab = createBottomTabNavigator();
 
 export const HomeTabNav = (): JSX.Element => {
   return (
     <Tab.Navigator>
-      {/* Feed */}
+      {/* Communities */}
       <Tab.Screen
         name={ScreenNames.Communities}
         component={CommunitiesScreen}
         options={{
           // headerShown: false,
-          header: (props) => <CommunityListHeader />,
+          header: props => <CommunityListHeader />,
           tabBarIcon: props => (
             <MaterialIcon
               name={MaterialIconNames.Home}
@@ -32,11 +33,15 @@ export const HomeTabNav = (): JSX.Element => {
           tabBarShowLabel: false,
         }}
       />
+      {/* Main Menu */}
       <Tab.Screen
-        name={ScreenNames.Login}
-        component={LoginScreen}
+        name={ScreenNames.MainMenu}
+        component={MainMenuScreen}
         options={{
-          headerShown: false,
+          headerShown: true,
+          header(props) {
+            return <AccountSwitcher />;
+          },
           tabBarIcon: props => (
             <MaterialIcon
               name={MaterialIconNames.Menu}