+++ src/types/index.ts
@@ -23,13 +23,23 @@ export enum MaterialIconNames {
   CheckCircleOutline = 'check-circle-outline',
   Home = 'home',
   Menu = 'menu',
+  MenuUp = 'menu-up',
+  MenuDown = 'menu-down',
   MinusCircleOutline = 'minus-circle-outline',
   PlusCircleOutline = 'plus-circle-outline',
+}
+
+export enum UnicodeText {
+  Bullet = '\u2022',
+}
+
+export enum ScreenMargin {
+  Horizontal = 20,
+  Vertical = 20,
 }
 
 /** Size for the tab icons in the bottom tab navigator */
 export const TAB_ICON_SIZE = 40;
 
-export enum UnicodeText {
-  Bullet = '\u2022',
-}
+/** Default timeout for api calls in milliseconds */
+export const DEFAULT_API_TIMEOUT = 5000;