+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - #3: Add RN Paper
 - #1: Add Typedoc support
 - #7: Add Redux
+- #13: Add Login Screen/Form